Put slices and selectors into separate files
