+++ src/features/filters/filtersSelectors.js
@@ -1,0 +1,2 @@
+export const selectDate = state => state.filters.selectedDate;
+export const selectFilter = state => state.filters.selectedFilter;
