+++ src/features/todos/todosSlice.js
@@ -1,5 +1,4 @@
 import { createSlice, createEntityAdapter } from "@reduxjs/toolkit";
-import { createSelector } from 'reselect';
 
 // Normalising state using createEntityAdapter
 const todoAdapter = createEntityAdapter();
@@ -38,20 +37,6 @@ Todo entry structure:
     date,      - contains the todo's deadline (string) - OPTIONAL
 }
 */
-
-// A factory function for a memoized selector for todos
-// The input reducer returns a stable reference (no change in state means no reference)
-// Gets an object of todos from state, converts it into an array, filters the array based on completion status
-export const makeSelectTodos = (completed) => createSelector(
-    state => state.todos.entities,
-    todos => {
-        const todosList = Object.values(todos);
-        return todosList.filter(todo => todo.completed === completed);
-    },
-);
-
-// Don't need memoization here, just a simple selector
-export const selectTodo = id => state => state.todos.entities[id];
 
 // Export actions
 export const { addTodo, deleteTodo, toggleTodo, updateTodo } = todosSlice.actions;
